fastapi/requirements.txt
fastapi
uvicorn
opentelemetry-sdk
opentelemetry-exporter-zipkin
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests